requirements.txt
fpdf==1.7.2
Pillow==7.0.0
